fix: consolidate typed worktree contract files
