+++ server/core/repo_boundary_test.go
@@ -324,7 +324,6 @@ func TestSharedClientUIRemainsDTOOnly(t *testing.T) {
 		"WorktreeTransitionID.UnmarshalJSON":                      {},
 		"WorktreeTransitionID.Validate":                           {},
 		"WorktreeTransitionOutcome.Validate":                      {},
-		"WorktreeDirtyState.Validate":                             {},
 		"validateHydrationBackgrounds":                            {},
 		"validateHydrationPrompts":                                {},
 		"validateHydrationQueuedMessages":                         {},
