+++ shared/clientui/worktree_cleanliness.go
@@ -17,7 +17,3 @@ type WorktreeDirtyState struct {
 	DirtyFileCount *int                   `json:"dirty_file_count,omitempty"`
 	UnknownCause   *string                `json:"unknown_cause,omitempty"`
 }
-
-func (state WorktreeDirtyState) Validate() error {
-	return worktreecontract.ValidateDirtyState(state.Kind, state.DirtyFileCount, state.UnknownCause)
-}
